requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
djangorestframework==3.16.1
Faker==22.0.0
idna==3.10
iniconfig==2.1.0
Markdown==3.8.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
